+++ Restaurant/app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name=".menuControlActivity"></activity>
-        <activity android:name=".cornerActivity" />
+        <activity android:name=".corner_activity" />
         <activity android:name=".AddActivity" />
         <activity android:name=".ManagementActivity" />
         <activity android:name=".loginActivity" />
