+++ Restaurant/app/src/main/res/layout/corner_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".cornerActivity">
+    tools:context=".corner_activity">
 
     <ListView
         android:id="@+id/listView"
